// talanta/src/app/models/Player.ts
export class Player {
  playerFirstName: string;
  playerLastName: string;
  parentFirstName: string;
  parentLastName: string;
  parentPhone: number;
  playerHeight: number;
  playerWeight: number;
  playerAge: number;
  playerHealthStatus: string;
  playerSchool: string;
  playerResidence: string;
  playerADM: number;
  // password: string;
  _id: string;
}